Initial monster data for 13 True Ways and Bestiary.
New monster data, based on the master list created by Sean Dunstan,
and Detective Clayton (and thanks to Cat Tobin for providing the link
to this).

The new data includes only basic tagging, so human intervention will
be required for adding additional sensible tags to those monsters.

This data also includes page numbers. This information will be added
into the existing monster entries too and then should be supported in
the app.

Finally, this checkin includes src/py/... which is python scripts
for processing data, and src/data/... which is raw data for use in
generating bestiarys