Starting to build out the results view in react
The second to last step in getting the UI ported to react is getting the results generating react components.
This is the first step to doing that. Once complete the toplevel ui component can be ported to react too, and then
a cleaner model of data gathering and handling can be brought in