Step one of moving local storage access
Local storage access has to occur on the main thread. The current implementation uses a LightWeightWorker to wrap it up.

Going forward this will be wrapped up with a MessageChannel which is then linked with the data access worker so that
the worker can (indirectly) access local storage and unify all the data access messaging.

This commit has the wrapped local storage but does not remove the party worker or take advantage of the local storage wrapper within the data worker.