Introduces the iter namespace
The iter namespace is utility functions that operate on iterable objects. These have been used to clean up some of the allocate function.

Now that all generation runs on a background process, and typescript supports downlevel iterators and generators, its time to start breaking up the crunchy allocate function into more managable pieces.

The previous implementation managed all the array walking and accumulation directly using index variables and pushing to lists so that it could be as snappy as possible for potentially long running generation on the UI thread to minimize the amount of time spent blocking. That is no longer a concern.

The generation is still clamped at 10,000 items. The ordering of results currently makes this hard to change (the results have to be groupBy'd rather than partitionBy'd so a concrete dataset has to be realised.

 * The groupMonsters function is now a simple paramteriser of the groupBy function.
 * allocation (within allocateMonsters) now generates repeats rather than allocating an array
 * for..of loop is used to consume the repeats