Introduces a forkable cursor for walking an array.
This change splits out the array walking and branching behaviour from the allocation algorithm. This is part of the work to clean up the algorithm now that it no longer needs to be quite so speed concious with the goal of making it easier to extend the rules and add flexibility and generation constraints.